Reorganized shared examples and helpers.
