+++ spec/grape/middleware/auth/basic_spec.rb
@@ -20,12 +20,12 @@ describe Grape::Middleware::Auth::Basic do
   end
   
   it 'should authenticate if given valid creds' do
-    get '/whatever', {}, 'HTTP_AUTHORIZATION' => encode_basic('admin','admin')
+    get '/whatever', {}, 'HTTP_AUTHORIZATION' => encode_basic_auth('admin','admin')
     last_response.status.should == 200
   end
   
   it 'should throw a 401 is wrong auth is given' do
-    get '/whatever', {}, 'HTTP_AUTHORIZATION' => encode_basic('admin','wrong')
+    get '/whatever', {}, 'HTTP_AUTHORIZATION' => encode_basic_auth('admin','wrong')
     last_response.status.should == 401
   end
 end
