+++ spec/spec_helper.rb
@@ -14,9 +14,6 @@ require 'rack/test'
 require 'pry'
 
 require 'base64'
-def encode_basic(username, password)
-  "Basic " + Base64.encode64("#{username}:#{password}")
-end
 
 require 'hashie/hash'
 
@@ -24,36 +21,7 @@ RSpec.configure do |config|
   config.include Rack::Test::Methods
 end
 
-# Versioning
-
-# Returns the path with options[:version] prefixed if options[:using] is :path.
-# Returns normal path otherwise.
-def versioned_path(options = {})
-  case options[:using]
-  when :path
-    File.join('/', options[:prefix] || '', options[:version], options[:path])
-  when :header
-    File.join('/', options[:prefix] || '', options[:path])
-  else
-    raise ArgumentError.new("unknown versioning strategy: #{options[:using]}")
-  end
+Dir["#{File.dirname(__FILE__)}/support/*.rb"].each do |file|
+  require file
 end
 
-def versioned_headers(options)
-  case options[:using]
-  when :path
-    {}  # no-op
-  when :header
-    {
-      'HTTP_ACCEPT' => "application/vnd.#{options[:vendor]}-#{options[:version]}+#{options[:format]}"
-    }
-  else
-    raise ArgumentError.new("unknown versioning strategy: #{options[:using]}")
-  end
-end
-
-def versioned_get(path, version_name, version_options = {})
-  path    = versioned_path(version_options.merge(:version => version_name, :path => path))
-  headers = versioned_headers(version_options.merge(:version => version_name))
-  get path, {}, headers
-end
