+++ spec/support/basic_auth_encode_helpers.rb
@@ -1,0 +1,4 @@
+def encode_basic_auth(username, password)
+  "Basic " + Base64.encode64("#{username}:#{password}")
+end
+
